+++ src/main/java/com/example/dncompany/service/admin/AdminUserService.java
@@ -107,7 +107,8 @@ public class AdminUserService {
     }
 
     @Transactional
-    public boolean banUser(Long userId, Integer banDays) {
+    // 벤 확인 처리를 위해 크기 비교 실패 했으면 0 false 반환 예정
+    public boolean banUser(Long userId, Long banDays) {
         return adminUserMapper.updateUserBanStatus(userId, banDays) > 0;
     }
 }
